backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
openai==0.28.1
python-multipart